project script/requirements.txt
pandas
xlwings
scipy
IPython
sklearn
matplotlib
plotly
dateparser
xlrd
pillow
cx_Oracle
xlwt
PyYAML